Fixed type cast

+++ wrtc/interfaces/media/tracks/audio_track_source.cpp
@@ -32,7 +32,7 @@ namespace wrtc {
             sink->OnData(
                     data.audioData.get(),
                     data.bitsPerSample,
-                    data.sampleRate,
+                    static_cast<int>(data.sampleRate),
                     data.channelCount,
                     data.numberOfFrames,
                     rtc::TimeMillis()
